Rename invalid-file.js to errors.js.
